Fix HealthKit Tests
